+++ locallib.php
@@ -2226,3 +2226,42 @@ function theme_boost_union_get_navbar_starredcoursespopover() {
 
     return $html;
 }
+
+/**
+ * Callback to add head elements.
+ * This function is implemented here and used from another location:
+ * -> function theme_boost_union_before_standard_html_head in lib.php (for releases up to Moodle 4.3)
+ *
+ * We use this callback
+ * -> to inject the flavour's CSS code to the page
+ * -> to add the touch icons to the page
+ *
+ * @return string The legacy implementation will return a string.
+ */
+function theme_boost_union_callbackimpl_before_standard_html() {
+    global $CFG, $PAGE;
+
+    // Require local library.
+    require_once($CFG->dirroot.'/theme/boost_union/locallib.php');
+
+    // Initialize HTML.
+    $html = '';
+
+    // If a theme other than Boost Union or a child theme of it is active, return directly.
+    // This is necessary as the callback is called regardless of the active theme.
+    if ($PAGE->theme->name != 'boost_union' && !in_array('boost_union', $PAGE->theme->parents)) {
+        return $html;
+    }
+
+    // Require local library.
+    require_once($CFG->dirroot . '/theme/boost_union/locallib.php');
+
+    // Add the flavour CSS to the page.
+    theme_boost_union_add_flavourcss_to_page();
+
+    // Add the touch icons to the page.
+    $html .= theme_boost_union_get_touchicons_html_for_page();
+
+    // Return the HTML code.
+    return $html;
+}